requirements.txt
youtube-transcript-api
google-generativeai
python-dotenv
fastapi
uvicorn
sentence-transformers
qdrant-client
fastapi
pydantic
uvicorn